all seed data added

+++ models/Decor.js
@@ -20,7 +20,7 @@ Decor.init(
         allowNull: false,
       },
     price: {
-      type: DataTypes.DECIMAL,
+      type: DataTypes.FLOAT,
       allowNull: false,
     },
     link: {
